Adding a basic client class
It uses the chosen interface.

When the main_menu issues a create or join game command, the method for
that is called. The former will create a server and connect to it, then
have it create a game with the specified scenario. The latter (which the
former then calls), will connect if need be, and then initiate a
GameEngine object using the scenario the server tells it to.

Inside GameEngine, the players, galaxy (map), resources, etc. will be
created as objects and passed around. These objects map transparently
through the connection to their corresponding location in the
database. This GameEngine will iterate through the turns of the game
until either paused or an end game condition is reached. At each turn,
it prompts the user for a course of action out of a list of available
actions (provided by the database); in this way, the user will never be
able to ask for an action the he or she cannot do. The user will be able
to move through different objects with different actions, and ends their
turn when all available action paths have been completed. Yay for state
machines.