Replacing dataSnag.py with data_parser function
dataSnag.py used three different functions to accomplish the same goal:
parsing the CSV data from the dat_files folder. data_parser replaces the
entirety of that module with one function that returns a list of lists
representing groups of data from the lines of the data file. It uses a
regular expression to capture and split this data properly.